bugs resolved and more settings added

+++ src/components/Solves/Solves.jsx
@@ -156,8 +156,27 @@ export default function Solves({
                   <p
                     className="solveDelete"
                     onClick={() => {
-                      handleDeleteSolve(solve);
-                      rearrangeSno();
+                      Swal.fire({
+                        title: "Are you sure?",
+                        text: "You won't be able to revert this!",
+                        icon: "warning",
+                        showCancelButton: true,
+                        confirmButtonColor: "#3085d6",
+                        cancelButtonColor: "#d33",
+                        confirmButtonText: "Yes, delete it!"
+                      }).then((result) => {
+                        if (result.isConfirmed) {
+                          Swal.fire({
+                            position: "center",
+                            icon: "success",
+                            title: "Your solve has been deleted",
+                            showConfirmButton: false,
+                            timer: 900
+                          });
+                          handleDeleteSolve(solve);
+                          rearrangeSno();
+                        }
+                      })
                     }}
                   >
                     X
